892 cleanup code
